refactor: split cluster timer packages
