+++ CHANGELOG.md
@@ -76,6 +76,7 @@
 
 ### Changed
 
+- **Cluster timer code now lives in domain subpackages.** Durable one-shot timers moved under `app.epistola.suite.cluster.timers`, durable recurring scheduled tasks moved under `app.epistola.suite.cluster.schedules`, and shared node registry/capability infrastructure remains in `app.epistola.suite.cluster`.
 - **UI request errors now return `application/problem+json`.** `UiExceptionFilter` emits a single structured envelope: any caller that accepts JSON (`application/json` or `application/problem+json`) or is an HTMX request gets an RFC 9457 problem body (`type`/`title`/`status`/`detail`/`instance`, `type` URI shared with the REST API) instead of falling through to Spring's whitelabel error page with a stacktrace; only true HTML navigations get a container error page. The legacy `{ "error": message }` envelope is gone — HTMX `responseError` handlers and the editor now read `detail` (with an `error` fallback for safety). Unknown exceptions stay an opaque 500. Known-exception statuses are unchanged (read-only catalog stays 403 on the UI surface). The **UI Spring Security chain** applies the same content negotiation to authentication (401) and access-denied (403, including CSRF rejections) failures — previously these fell through to Spring's whitelabel `/error` page for fetch/XHR callers; HTML navigations keep the login redirect / container error page. Remaining UI error gaps were closed: a new `UiHandlerExceptionResolver` returns problem+json for **pre-dispatch framework errors** (404 unknown path, 405, 415, 406) and **uploads over the limit** (413, new `PAYLOAD_TOO_LARGE` type) for structured callers; the **editor draft-save** route now returns `ValidationProblemDetail` problem+json (replacing the bespoke `{type,message,errors}` envelope) and answers a missing/blank `templateModel` or bad path with a 400 problem instead of a 500/empty body.
 - **Error discriminator is now the problem `type` URI, not a separate `code`.** The contract dropped the `code` member (RFC 9457 intends `type` to be the machine-readable discriminator, and `type` already encoded the code as its slug). The REST API stops emitting `code` on the wire (`ApiProblemType.code` stays as the internal slug source); framework-generated 4xx no longer need a `code` stamped on them. The editor draft-save route now returns the problem `type` URI instead of `code`, and the editor (`save-service`, `binding-errors`, `editor.html`) switches on the `type` URI (e.g. `https://epistola.app/errors/node-parameter-binding-syntax-invalid`). Wire `type`/`title`/`status`/`detail`/`instance`/`errors` are unchanged; only the redundant `code` field is gone. Requires `epistola-contract` ≥ the type-only release.
 - **Align error handling with the RFC 9457 `epistola-contract`.** The contract adopted RFC 9457 (which obsoletes RFC 7807, no wire-format change), mapped `ProblemDetail`/`ValidationProblemDetail` onto Spring's native `org.springframework.http.ProblemDetail`, and removed the deprecated generated `FieldError`/`ValidationErrorResponse` types. This app already used the native `ProblemDetail` as its internal builder, so the migration is type-only: field-level errors now use the contract's `ValidationError` (same `field`/`message`/`rejectedValue` shape) in `ProblemDetails`, `ApiExceptionHandler`, and the validation tests; the editor draft-save route's pre-Problem-Details envelope (`code`/`message`/`errors[]`) is now a local `DraftValidationErrorResponse` instead of the removed contract DTO. Wire JSON is byte-identical; ADR 0004 amended. Requires `epistola-contract` ≥ the RFC 9457 release.
