+++ modules/epistola-core/src/main/kotlin/app/epistola/suite/cluster/schedules/ClusterScheduledTaskDefinition.kt
@@ -1,5 +1,6 @@
-package app.epistola.suite.cluster
+package app.epistola.suite.cluster.schedules
 
+import app.epistola.suite.cluster.ClusterProperties
 import app.epistola.suite.common.ids.TenantKey
 
 data class ClusterScheduledTaskDefinition(
