+++ modules/epistola-core/src/main/kotlin/app/epistola/suite/cluster/schedules/ClusterScheduledTaskRegistry.kt
@@ -1,5 +1,6 @@
-package app.epistola.suite.cluster
+package app.epistola.suite.cluster.schedules
 
+import app.epistola.suite.cluster.ClusterProperties
 import app.epistola.suite.common.ids.TenantKey
 import app.epistola.suite.observability.NodeIdentity
 import org.jdbi.v3.core.Jdbi
