+++ modules/epistola-core/src/main/kotlin/app/epistola/suite/cluster/timers/ClusterTimer.kt
@@ -1,4 +1,4 @@
-package app.epistola.suite.cluster
+package app.epistola.suite.cluster.timers
 
 import app.epistola.suite.common.ids.TenantKey
 import java.time.OffsetDateTime
