+++ modules/epistola-core/src/main/kotlin/app/epistola/suite/cluster/timers/ClusterTimerHandler.kt
@@ -1,4 +1,4 @@
-package app.epistola.suite.cluster
+package app.epistola.suite.cluster.timers
 
 import java.time.OffsetDateTime
