+++ modules/epistola-core/src/main/kotlin/app/epistola/suite/cluster/timers/ClusterTimerScheduler.kt
@@ -1,6 +1,9 @@
-package app.epistola.suite.cluster
+package app.epistola.suite.cluster.timers
 
 import app.epistola.suite.background.BackgroundExecutionContext
+import app.epistola.suite.cluster.ClusterNode
+import app.epistola.suite.cluster.ClusterNodeRegistry
+import app.epistola.suite.cluster.ClusterProperties
 import app.epistola.suite.observability.NodeIdentity
 import app.epistola.suite.observability.recordScheduledTask
 import io.micrometer.core.instrument.MeterRegistry
