+++ modules/epistola-core/src/test/kotlin/app/epistola/suite/cluster/schedules/ClusterScheduledTaskRegistryIT.kt
@@ -1,5 +1,7 @@
-package app.epistola.suite.cluster
+package app.epistola.suite.cluster.schedules
 
+import app.epistola.suite.cluster.ClusterNodeRegistry
+import app.epistola.suite.cluster.ClusterProperties
 import app.epistola.suite.mediator.execute
 import app.epistola.suite.mediator.query
 import app.epistola.suite.observability.NodeIdentity
