+++ modules/epistola-core/src/test/kotlin/app/epistola/suite/cluster/schedules/ClusterScheduledTaskSchedulerIT.kt
@@ -1,4 +1,4 @@
-package app.epistola.suite.cluster
+package app.epistola.suite.cluster.schedules
 
 import app.epistola.suite.testing.IntegrationTestBase
 import org.assertj.core.api.Assertions.assertThat
