+++ modules/epistola-core/src/test/kotlin/app/epistola/suite/generation/collect/maintenance/StaleConsumerNodeReaperIT.kt
@@ -1,6 +1,6 @@
 package app.epistola.suite.generation.collect.maintenance
 
-import app.epistola.suite.cluster.ClusterScheduledTaskRegistry
+import app.epistola.suite.cluster.schedules.ClusterScheduledTaskRegistry
 import app.epistola.suite.common.ids.TenantKey
 import app.epistola.suite.testing.IntegrationTestBase
 import org.assertj.core.api.Assertions.assertThat
